uniq/requirements.txt
Django==1.6.6
django-filter==0.7
djangorestframework==2.4.0
mongoengine==0.8.7
pymongo==2.7.1
